moved thread from terminal creation to instance initialization
